python/requirements.txt
pyserial>=3.5
python-dotenv>=1.0.0
RPi.GPIO>=0.7.1

